test: remove parent class SymfonyConstraints
